refactor(UserInterface): reorganize folder layout for views
- put user interface relating source files into userinterface package
