+++ src/main/java/racingcar/view/userinferface/printer/ConsolePrinter.java
@@ -1,4 +1,4 @@
-package racingcar.view.printer;
+package racingcar.view.userinferface.printer;
 
 public class ConsolePrinter implements Printer {
