+++ src/main/java/racingcar/view/userinferface/printer/Printer.java
@@ -1,4 +1,4 @@
-package racingcar.view.printer;
+package racingcar.view.userinferface.printer;
 
 public interface Printer {
     void print(String s);
