+++ src/main/java/racingcar/view/userinferface/scanner/Scanner.java
@@ -1,4 +1,4 @@
-package racingcar.view.scanner;
+package racingcar.view.userinferface.scanner;
 
 public interface Scanner {
     String read();
